requirements.txt
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
ephem==4.1.5
Flask==3.0.0
fonttools==4.43.1
geographiclib==2.0
geopy==2.4.0
h3==3.7.6
itsdangerous==2.1.2
Jinja2==3.1.2
jplephem==2.19
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pycparser==2.21
pyparsing==3.1.1
pyswisseph==2.10.3.2
python-dateutil==2.8.2
pytz==2023.3.post1
setuptools==69.0.2
sgp4==2.22
six==1.16.0
skyfield==1.46
timezonefinder==6.2.0
Werkzeug==3.0.1
